faster search logic
for the search logic, i realized if it works for size, then it also works for size+1, size+2, etc.
thus, with the new bounds logic we can iterate in reverse order and stop on the first failure.
we just have to record a bit of extra state information due to inconvenient search order.
but if min is n, this eliminates the need to investigate all combinations of 1, 2, 3, ..., n-1 elements.
so this is increasingly faster for larger problems, which we're now getting into.