branch and bound
changed the search function to be branch and bound.
for tested input sizes this is considerably faster.
only major user-level difference is that rows x cols (input) now reshapes to have more rows than columns.
this is because the new branch and bound logic searches in row major order and the bound function gets better with increasing number of rows to validate.