arbitrary tessellations
added support for performing arbitrary tessellations.
i'd considered doing this from the start but thought the automatic phase shifting would be tricky.
but because some shapes have weird tessellation basis vectors, i've decided to not even try to calculate them.

the input for the new 'geo' search option is a geometry file.
first line is 4 integers representing two tessellation basis vectors.
the rest of the file is a binary histogram where 1's denote in the geometry and 0 not in.

the geometry can be literally anything, including both convex and concave shapes.
it doesn't even have to be contiguous: you could put a hole in a diamond and as long as it's somehow filled in uniquely via tessellation you're good to go.
that being said, there are some limitations with the current tessellation mapping algorithm (which would only be a problem for very weird shapes).
for instance, it only attempts to perform a 5x5 tessellation.
this might be too small or too much, so in the future this will be improved to be faster and work for every possible type of shape.

from there it simply runs through every possible combination of old set contents and checks if they work.

CURRENT_BEST is removed and replaced by a trailing (required) command line argument.

added an Adjacent4 iterator for getting geometry boundaries, but i ended up not using it - will likely be removed in a future commit.

added 2 example geometry files for reconstructing some known 1/4 diamond solutions.