more old types and graphs
previously, only supported looking for standard OLD sets.
now supports OLD and DET:OLD (detector problem).
with how it's currently written it could also easily be extended to other related problems.

additionally, used to only support king grid (hence the github repo name).
now supports king grid and triangular grid.

these are specified by a new command line argument of the form e.g. det:tri or old:king.

temporarily removed the rand iteration feature.
it's just not very useful because either the problem is small enough that exhaustive is doable or it's so large that random iteration is a negligible chance of success.
may be added back later, but i'm not prioretizing its reimplementation in the new system.

eternal sadness: i had to use an unsafe block to get the borrow checker off my back :(
it's a struct field that's owned but never mutated after construction.
for some reason borrow checker isn't able to deduce that calling a &mut self method while that field is borrowed immutably is safe.
i'll try to rewrite it in such a way that no unsafe code is used later on, but it's perfectly safe as is.