cleanup
did a little bit of code reorganizing to be simpler, safer, and have less code dupe:

refactored RectSolver into RectSolver and RectSolverBase.
RectSolver is the Solver, and base has everything except phases.
this allows us to borrow phases immutably while performing mutation on the rest.
so the (single) unsafe block could now be removed.

added a Tessellation trait.
apparently rust doesn't support generic associated types, so this is not implemented *exactly* how i envisioned:
has a try_satisfy<Codes, Adj>(thresh) function that is equivalent to getting an appropriate solver and running the appropriate try_satisfy() on it.

rewrote all the stuff in main() after constructing rect/geo to refer to a helper function.
the helper, of course, is generic on Tessellation and parses/applies the mode string.
because we also need to print the tessellation on success, it has to handle the result as well, so main() has nothing left to do afterwards.
i would've liked this to be a generic lambda or a Box<dyn Tessellation> but rust doesn't support either of those.