minor geo change
during the last refactor i removed the part of geo logic that printed the loaded geometry immediately (blank) just so user could verify shape and basis vectors.
i thought it was unnecessary, but now that i'm using it in batches i find it would in fact be useful.
added back now as before.
unnecessary for rect due to rect simplicity.