added theo exec mode
added a new execution mode: theo.
the other modes all try to find configurations of vertices that yield solutions, which act as inclusing upper bounds.
the theo option istead generates an exhaustive share argument to prove an inclusive lower bound.
using these two modes in tandem could theoretically produce and prove that some configuration is optimal.
out of the box works for OLD and DET:OLD sets on the infinite grid, triangular grid, or king grid.

the current algorithm it uses is quite simple.
it simply tests every possible vertex domination configuration and calculates the maximum domination share.
the lower bound is then simply the reciprocal.

while this works, it is not as good as it could be.
to form a lower bound on an infinite graph we instead want the maximum average share, not maximum share.
so this always gives correct values, but they're not as tight as they could be with averaging logic.

hopefully, this will be added soon.