area expand bug fix
fixed a bug in the new theo logic where it was generating the iteration bounds incorrectly.
it was adding distance 2 positions, but wasn't adding the pre-existing radius 1 positions.
fixed by simply cloning and extending area instead, as the tessellation area constructors now do.

fixed a similar bug in the tessellation area expansion logic.
however, as it was using Adjacent8 rather than the Adjacent4 case that was causing a problem for theo, this was only an issue for expanding a 1x1 vertex geometry.
this was never tested before, but also wasn't ever used, so all pre-existing geo tessellation analysis is still valid.