added finite graphs
added the ability to find old, red:old, and det:old sets on finite graphs.
these graphs are defined by a graph file, which is a list of whitespace separated tokens, each of which is a colon-separated pair of vertex labels representing an undirected edge.
this mode is accessible via the new "finite" execution mode.
aside from the file path and old-variant type, it takes one command line arg denoting the size of the detector set to search for.