implemented theo thresh arg
previously theo had a required thresh arg representing density.
it was completely unused.
now represents a thresh value (>= 1).
is used to conditionally pipe problematic configurations to a handler function.
handler function is a new arg to theo solver calc method (stored as optional dyn).
if share exceeds thresh, calls handler - if present - with current configuration.
currently the handler theo_helper() uses just prints the share and configuration to the screen so the user can see what configs resulted in a high share value.

removed the prevs stack from geo solver.
it was added with the intent to store extra information that could be used for branch/bound, but i never came up with a better pruning algorithm so it's useless since all we need is the len, which is identical to old_set len.