hex grid
added hex grid iterator and gave it rect/geo/theo impl for everything.

rect mode no longer enforces rows>=cols by conditional transpose.
this logic is sound for some grid types, but not all of them.
thus from now on we'll take the conservative side and just do exactly what user requests.
not sure if it'd be a problem for any grid type currently included (since they're all regular), but it would eventually be a problem down the road.