ic theo bug fix
found a bug in the new identifying code theo logic.
problem was caused by having duplicate positions in the adjacent iterator.
we were constructing a closed neighborhood by explicitly chaining the verted position with the adjacent iterator, which was assumed to be open.
thus, when given a closed iterator it has 2 copies of the same center position.

this is fixed by adding a new closed_neighborhood_unord() method to the AdjacentIterator trait.
for closed iterator types this is emplemented by returning itself.
for open iterator types it returns both itself and the center position, connected as a chaind iterator.
the position order for this function is unordered, hence the name - this allows for high speed.

added OpenIterator and ClosedIterator unit traits.
they are implemented appropriately for all the various adjacent iterator types for use in advanced trait bounds in the even we require open/closed only.