theo refactor
i have a plan to add averaging share arguments to theo, which will make its lower bound value VASTLY superior to what it currently does.
in preparation, i've put together this refactor which is functionally identical to the previous commit but is restructured in a way that will make it relatively easy to add averaging arguments from here.

AdjacentIterator now implements Clone.
added AdjacentIterator::at() which is like new() but takes a tuple (point).

added the Geometry type.
this has a shape and detectors and is able to print itself prettily.
this is used to implement GeometryTessellation.
is also used for fancy theo printing, which is much more readable for non-gridish graphs.