added averaging arguments
vastly improved the power (at the expense of a LOT of speed) of the theo lower bound searcher.
it now (intelligently) expands up to radius 4 (instead of 2) and uses this information to construct averaging arguments.
its arguments aren't the strongest possible, but they suffice for reducing the number of problem cases.
'averagees' must have share under thresh and their open neighbors must as well.
this is a conservative decision that at least ensures that any average share arguments it constructs are sound (though a human could potentially do better).

the new algorithm tries to expand as little of the state space as possible while still being able to construct powerful averaging arguments.
to facilitate this, there are many MANY short circuiting points in the logic.

codesets now have an associated constant MIN_DOM representing the minimum domination count for any valid vertex.
for most things this is 2.
this is used as one of the bounding functions for new theo logic.