added entropy search mode
added entropy-rect and entropy-geo.
these take a shape (rect or geo) and an entropy size.
iterates over all possible geometries of that size which fit inside the big rect/geo.
for sufficiently large starting regions, this is equivalent to iterating over all shapes of a given size.
they are not required to be contiguous.

added subgeometries() helper function to geometry to iterate over all subgeometries of a given size.

moved raw geometry file reading to geometry type.
no longer needs to be in geometry tessellation stuff.