LOC refactor + LD lower bounds
added LOC trait to codesets.
this is used to represent locating codes for various dominating sets.
added RegularLOC for most types and LDLOC for LD sets.
the only difference so far is that LOC::dom() behaves differently for LDLOC (adds 1 if detector).
LDLOC expects open neighborhood of codes and uses the is_detector flag to represent self-domination.

the codesets::Set type no longer has MIN_DOM (was not needed in only place it was used).
this is in the hopes of eventually adding generic redundant sets, which would require member variables rather than constants.
codesets::Set is also no longer a generic - it's a normal trait with an associated type.
this is to remove ambiguity issues, but it makes more sense this way anyway.
LOC is similarly not generic.
each Set has an associated Item type, and an associated LocatingCode type which implements LOC<Item>.

all locating code logic is changed to return some LOC type specific to the code set being used.
this includes the dom() count used in lower bound searcher, which fixes one of the 2 problems for ld lower bounds.
the other problem was that lower bound was using the same adj iterator type for locating codes and share - ld needs open codes but is closed domination, which requires closed share.
thus lower bound now takes 2 adj iterator types, one for codes/graph structure, and another for share neighborhood.
they are assumed to be open/closed variants of one another, though this is not enforced.
these 2 changes effectively fix ld lower bound (which was removed due to failing unit tests), so it's added back in now.

redid all the lower bound unit tests to be in terms of the theo_helper() function, which is much more readable.