added det:ld set
error-detecting LD sets previously didn't have a characteristic (just the definition), and so weren't practical to include in this program.
however, i've recently proven a characteristic for them, so now it's super easy, barely an inconvenience.
added codesets::DETLD to represent a code set with my characteristic's constraints.
added a unit test for it cause it's at this point the most complicated detection system covered by this program.

wrapped symmetric difference (size) logic into its own function to be more readable.
added a unit test because this is rust and i'd be crazy not to.
did the same for sharp difference (size) (i.e. max length minus equal) for composing (symmetric) sharp distinguishing constraints like DET:OLD and DET:IC.