rational theo update
the logic used in theo assumes flawless rational numbers.
the workaround i was using was to slightly overshoot the thresh value by a small amount.
however, with the new discharging mode that has become problematic as well.
so now i've imported the num crate and have replaced all theo stuff with BigRational.
this is currently type aliased as Share because i was experimenting with Ratio<i32> and Ratio<i64>, but ultimately resorted to Ratio<BigInt> for safety.
in the future, this type alias will likely be removed.

theo now takes as argument the lower bound itself, rather than the share.
this is converted to a rational number.
input is of form "{p}/{q}" or "0.{f}".
the first form is done for easy testing of e.g. is 4/13 a lower bound?
the second form is so we can do decimal inputs in the density space for binary searching for the rational boundary point.

added a short little python script which takes an approx floating point value and a thresh.
it then outputs a rational representation of the value which is within thresh of the given value.

with the new way avg works (as well as discharging, which itself is new) the inverse of highest avg share is no longer guaranteed to be a lower bound (cause we sometimes undershoot, but that was only possible due to the specific choice of target bound).
thus, theo now simply returns bool denoting if it succeeded or not.
on success, prints the bound, otherwise prints problems and nothing else.
this also means we don't even need to keep track of max avg share anymore.

updated unit tests for theo stuff - will be expanded later.
fixed a bug where theo was always returning true if pipe is None (because if no pipe it was skipping problem generation entirely).
additionally, now theo when passed None for the pipe will stop on the first problem.
this essentially allows us to short-circuit failure if we don't care about the problems.