added red:ld set
added the REDLD codeset type that has the same constraints as just derived in our most recent paper for RED:LD sets.
the constraints were a bit too complicated for the pre-existing interface, as it requires knowing the position of all detectors.
thus, i've changed the interface slightly to add the current position as a requirement for codesets::LOC::new().
most implementers ignore this, but the new REDLOC type stores it as well as everything else.