auto-theo execution modes
added the auto-theo execution modes which do a binary search in the density space for the largest upperbound we can prove.
added a utility function util::rationalize() which takes a value and thresh and returns the simplest rational number in the bounds.
this is basically the python code but better because it's done losslessly instead of using floating point like python version does.
also, improved both versions: we don't need to look at floor and ceil - we want the closest, which is always given by simply rounding.