improved theo
improved theo searching in some cases.
some cases esp. tmb graph were getting sub-optimal max shares, which lead discharging/averaging to fail when it could have succeeded.
this never results in wrong answers, but is potentially sub-optimal.
now fixes via a hack where we recompute share a different way in the event that avg/dis fails for some case; however, this is twice as slow and potentially still sub-optimal.
i have a different idea to facilitate the exact desired behavior, but this commit is left for revisioning and history in case its code is ever needed again (cause it will probably be removed+replaced on next commit).