simplified better theo logic
as mentioned in previous commit, this commit is basically the same thing but better.
turns out, it's sufficient - and indeed strictly better - to just surround the expanded space in detectors and test validity on the entire closure.
this avoids a lot of unnecessary work and we no longer need to have special functions for deep dom-counting, which has no ability to short circuit.
now simplified to just one expansion field and an exterior set for both neighborlands and borderlands.
speed-wise this is a bit slower than last commit, but is more powerful.
additionally, this is a trivial implementation of the new expansion logic - breaking some large areas into smaller shells would likely make it even faster than before while still being stronger.

this commit leaves a few things that can be removed later - is mostly just here to save the progress, as everything has been tested and is definitely working.