parallel entropy
changed entropy to use parallel searching.
for any given geometry, the search can't really be parallelized easily.
but entropy considers every possible sub-geometry, and that part can be trivially parallelized with a consumer pattern.
it now does the same thing but speedy zoom zoom, not that that really helps the asymptotic complexity.
however, for reasonable inputs, this is a significant improvement (indeed, benchmarking shows linear speedups).

changed the parameter (set) type and graph type to be enums which impl FromStr.
this means we don't have to parse them multiple times during entropy geometry iteration.
additionally, they are now case-insensitive and some support multiple names for the same thing.