added smart-entropy script
the entropy search mode of the main program is very powerful, but it only takes the size of entropy.
for many sizes we might test (1 exec of main program for each size) some sizes may be provably impossible to be better based on the proven lower bound.
smart-entropy.py takes the target value (typically the current best upper bound minus epsilon) and the lower bound (due to rounding errors, typically current best lower bound minus epsilon).
it finds all the sizes that could conceivably work and shows what values they could lead to.
this tremendously reduces the number of entropy sizes (for arbitrary bounding regions) must be tested as a whole if done exhaustively (as i typically do).