bug fix + cleanup
found a bug where the upper bound searchers (both rect and geo) could be fooled into giving a false positive for irregular graphs due to not testing all the induced classes.
we did and still assume in generality the root class to be 0, but it wasn't checking the others.
this is ok if there is only one class or if the number of induced classes happens to be 1, but otherwise is a problem.
this came up for rect 4 6 ic hex 5/12, which i manually verified to be an incorrect solution.
to remedy this, the is_old() function (last step of solver) now checks all induced vertex classes.
the current impl is O(1) for any class we don't induce, so it's practically a free patch.

removed rect stuff entirely - all of it - for now and forever.
the rect stuff was fast, yes, but overly complicated.
it was always hard for even me to modify it, and i wrote the blasted thing.
not to mention the phase pseudovector has opposite directions than the geo stuff, which is confusing.
so i've elected in this commit to remove it entirely.
the rect exec mode still exists, but is defined as using the geo logic on Geometry::rectangle(rows, cols).
it still has the 1xn nx1 restriction, but that will likely be removed in the near future.
this commit was primarily a bug fix.

to fix the class logic, AdjacentIterator now has additional functions:
class(row, col) - gets the class (usize) of specified position.
with_class(row, col, class) - same as new(row, col) but uses the given class.
new(row, col) is now simply defined to be with_class(row, col, class(row, col)).
at(p) remains an alias for new(p.0, p.1).
with_class() is used by the new locating code logic, which requires a (pseudo) offset to apply.
by this i mean it uses the same positions but pretends to be somewhere else in terms of graph connectivity (i.e. different class).