fixed some clippy warnings
fixed a few clippy warnings.
some of them were good, but others i thought were far too pedantic to care about.

changed the way PointMap::iter() works so that it doesn't need to do division operations.
this appears to be a bit faster, but it might just be because of removing a wrapper layer with filter_map.