stronger discharging logic
my discharging technique technically works for diffusion factor k being the maximum number of simultaneous influx nodes.
previously, we were using the more conservative k being non-simultanous, as this is much faster to compute.
i thought the full logic would be too slow, but apparently it's not such a big deal.
the field unions are not as disjoint as i presumed they would be (without testing), being at worst on the order of first-order expansion.
so, if implemented correctly, this is as strong as this discharging technique can be.